Fix crash when taking screenshots from overview
When taking screenshots from overview, the current screenshot
wasn't being cleared, leading to a crash when we tried to add the
view to the window again. Clearing the screenshot first (as we do
when taking screenshots from hardware/global actions) fixes the
problem.

Fixes: 151153057
Test: manual (verified that error occurs if you tap the screenshot
button while the UI is up pre-fix, and crash does not happen
post-fix)

Change-Id: Ia33344ce19c912241e18d3053ccdc51845603ca3

Former-commit-id: ab359beb2be72d0d6318f01fe6408946d4d162f3