[RESTRICT AUTOMERGE] Allow bar to be transparent during animating.
When top window request fullscreen and device contain a real cutout
and the layoutInDisplayCutout is default or never, a letterbox will
be created for this activity, which cause
BarController#isTransparentAllowed return false so the status bar
cannot be transparent, the bar will shows a black background
during hidding animation.
A simple workaround is to allow the bar can be transparent
during hidding, since the surface of letterbox is not really
overlapped with bar because it's also animating.

Bug: 150048810
Test: No black background shows on status bar when opening
fullscreen app.
Test: atest DisplayPolicyTests

Change-Id: I2dd5b1fbad8e79ec80f799d63bba8824acc6a639

Former-commit-id: 44f50d484de99b0801649ac052acd46c7be66192