Raise tz standards for uninitialized devices
Removes the lowering of standards applied for tz detection on
uninitialized devices. Before this change, the device would accept any
telephony tz suggestion, which could include lower-quality answers; i.e.
those where the zone was chosen based on a country's default zone, even
when there are multiple different zones to choose from.

Unfortunately, the stock device setup wizard assumes any time zone
change made while it is running is a high quality answer and so it then
doesn't ask the user to confirm the choice. Without an API it has no way
of telling whether the time zone chosen was a high or low quality
answer. This should eventually be resolved with improved system APIs.

Bug: 143721631
Test: atest services/tests/servicestests/src/com/android/server/timezonedetector
Change-Id: I1dbd1046f0e641e609ecc8809cf86b7ddb6111b3

Former-commit-id: ea6b34da8d4291390649f1ced7bd39c199ec6f9e