Send notification when reusing a stack as a task.
We used to always create a task even though we knew there was only one
task ever created under a stack. Thus the onTaskCreated() notification
is always sent with correct component name. However after we merged
stack & task in some cases, stack can be reused as a task and listeners
(e.g. SysUI) lost a useful notification as when stack is created
componentName is still null. Therefore send out the notification when
reuse is decided. Listeners can determine if it's a root task or leaf
task from the nullness of component name.

Bug: 150252489
Test: Smoke tests & TH.
Change-Id: I0896307fd6bbfa53d95843043235e506e1a97a43

Former-commit-id: a7c9f7cb4cc7dc4d4d93d543169523bcd9adcfb5