Add getAllNotifs() to CommonNotifCollection
The following CL needs to be able to iterate through all current
notifications, so we need this method.

Adds a method, getAllNotifs() to CommonNotifCollection. This means that
both NotifCollection and NotificationEntryManager must now implement
this method.

NotifCollection previously had the (confusingly-named)
getActiveNotifs(). Now renamed to getAllNotifs().

NotificationEntryManager previously had
getPendingAndActiveNotifications(), which wasn't very efficient.
Restructured to the class to not need to rebuild the collection every
time it was called. Moved the (one) existing client over to use it.

Bug: 112656837
Test: atest SystemUITests
Change-Id: Ic2d5860e080fff8cb11c37ccc1a92f0af761caab

Former-commit-id: 88aeaf7dc8a4f1df9926ed87194698b596889c0e