PIP: Add mouse hover/exit menu show support.
This adds the ability to show/hide menu based on hover enter/exit
events.

This also changes the InputConsumer so that it's now attached to the
pinned RootTask (Stack)'s SurfaceControl, so that the PipMenuActivity
appearing will not cause a new HOVER_EXIT + HOVER_ENTER pair event.

Bug: 144184591
Test: atest PipTests
Hover in/out of PIP activity, see menu show/hide.

Change-Id: I79536eb52cd234a82cbb7cb3004e31ada3ca0433

Former-commit-id: 34b6fbf7e6143f34274e664ad8139f678894199e