TouchExplorerTest: insure that no-change move events don't crash the system.
A device will sometimes generate move events that don't change coordinates from the previous event.
We need to be sure that any calculations the system performs can handle zero-change values so we now test that in TouchExplorerTest.
Specifically, we test it in both the touch exploring and dragging states.
Also improved error reporting in MotionEventMatcher to report motion event actions as strings rather than numbers.

Test: atest FrameworksServicesTests:TouchExplorerTest
Change-Id: Id5b551327b54619fcf07f76c503a78e3d5059041

Change-Id: Ic36b80f8eecc3640092ec0a5cf893631d120f6c6

Former-commit-id: 3118dae54a34182c556dcd6609712333a9d35663