Use task bounds as local animation frame in fixed rotation
When switching to a fixed rotation transformed activity, the activity
will be rotated but display and task are still in original rotation.
So the local animation should use the same coordinate space as display.

The insets for animation are not handled because aspect-scaled-thumbnail
animation is no longer used (replaced by recents animation). So the
methods in rotator are also simplified.

Bug: 151709552
Test: DisplayContentTests#testApplyTopFixedRotationTransform
Test: Enable fixed_rotation_transform and launch activity with
      ActivityOptions#makeClipRevealAnimation in different orientation.

Change-Id: I7c924719eb535252509305cd842cdb4e0279b858

Former-commit-id: f41034c72a8b66bf06a0837ae6a4126a259ebb6e