Disable deep press when long press is long
If long press timeout is not 'short', we disable deep press.

Also InputManagerService.java will now be responsible for keeping track
of the feature state.

In ViewConfiguration, we update the default value to 400 to match the
value in the settings (b/30159825)

Bug: 148311342
Bug: 30159825
Test: see the description in the frameworks/native change
Change-Id: I88b933e9e863d40e383afdc990e09b848e23192e
Merged-In: I88b933e9e863d40e383afdc990e09b848e23192e

Former-commit-id: ab9efea9359bea1533b12daa37186f0ec39a3e94