Improved the transitions and layout of the app name in conversations
The transition was using getTop instead of absolute positions which
didn't work well for the conversation layout.
Also, we were fading in views that were GONE and making them VISIBLE
which lead to bugs during the transitions.
Additionally, the header text would become visible when a notification
was removed from a group and reset.

Bug: 150905003
Test: visual, add conversations in groups and non-groups
Change-Id: I3c31006a1fc79f7d58cc1dd3d5af44129c9f02bb

Former-commit-id: cef53f36debad86406abb6a0bf87a1515a98923e