Add (hidden) fixupAppDir() API.
On devices without sdcardfs, filesystem permissions aren't magically
fixed up. This is problematic mostly in application-private directories,
which are owned by the application itself; if another process with
elevated permissions creates a file in these directories, the UID will
be wrong, and the owning package won't be able to access the files.

Allow processes that modify application-private directories (eg
DownloadManager) to ask the framework to fix up permissions on their
behalf.

Bug: 151986170
Test: WIP
Change-Id: Ic30f708d25d7702ec3a4569ee6ddf100732d89d1

Former-commit-id: 0f522bfebf43c214ccf0d940a810058e9f8c0f2c