Remove usage of sleep tokens during dreams
Previously, a sleep token wass acquired when the dream starts. This was
done because the dream was implemented as a floating window.
Acquiring the sleep token puts all other activities in a sleeping state.

Since ag/10493254, the dream is implemented using an activity. When the
activity shows on top of other activities, the rest are automatically
moved to a STOPPED state. Hence, the sleep token is no longer needed and
can be removed.

Bug: 152284980

Test: m && flash && verify dream start/stop works + it doesn't interfere
      with the keyguard
Change-Id: I00fc4b0b0e63702086fb283329605556936ec084

Former-commit-id: a37fc59fde30854a823ccd54ba0d0eaa6fea65f7