Correct the dirty region of magnifification bounds
The window type of window magnification is non-magnified. When
window magnification and fullscreen magnification are both enabled
during the test, we can find that the bounds of non-magnified area have
afterimage remained on the screen.
When drawing magnification bounds, Accessibility controller would clean
the dirty region of the canvas first then draw new magnification bounds
on the canvas. While the dirty region is always the current rectangle
of magnified region that not includes the one of the previous frame.
So the part of the previous frame would not be fully cleared. Hence the
afterimage would happen.
To fix it, we also have to correct the dirty region by the XOR of
magnified region and the old magnified region.

Bug: 152264396
Test: manual test
Change-Id: I19c796bdcf9eb80df8c05a83ed7a79e7981e77b1

Former-commit-id: 64900394f50c56d129979a35712772f25a023593