Ignore specified animation types when checking isAnimating
The check isAnimating is used in a lot of places to determine if the
apps can hide, show, removed, etc. However, we don't always want to
check if any animation is running, just if a specific animation.
This change adds a new method that allows the caller to check if
isAnmating but ignore certain animationTypes.

This is specifically needed for ScreenRotationAnimation refactor to
ensure we can ignore screen rotation animation when checking
isAnimating. Otherwise, screen rotation will block hiding closing apps
if rotation occurs at the same time.

Test: WindowContainerTests
Bug: 152333373
Bug: 149490428
Change-Id: I922e3a5698f5db4dcc819a59ebfde2cd0301d7d2

Former-commit-id: 9177c77262fb190659ce0b031d7eb3696196d067