Expose ImageDecoder#createSource(byte[] (, int, int))
Bug: 152322291
Test: Ib50eba48e3557a0147b95dd6ce3e8589ab07c127

These methods are used in various places in the system, and the system
will not be able to access them as @hidden APIs once ImageDecoder is
part of a mainline module. They were originally hidden to discourage
clients from keeping the encoded data allocated in a large buffer,
especially for animated images. But this is already possible using a
ByteBuffer.

Change-Id: If5350d4f2ff88ac634e48b9224d70954fd312d30

Former-commit-id: 8060fa83283a6186cb09b52591ba31f041bf5db3