Disabled back gesture when quickswitching if needed
We allow quickswitching apps in different
orientations by touchiing in the same region on
the device. To avoid conflicting touches between
the swipe gesture and the back gesture, we disable
the back if the rotation of the swipe location and
rotation of current device do not match.

Fixes: 150250451
Test: Tested quickswitch manually with
test apps fixed to different rotations.
Ensured back only showed when rotation of
touch and display matched.

Change-Id: If3b4d15eb4b66ce688b91d44a2ec16b3610ecf0a

Former-commit-id: f848e9d439646b31f0d445b5a781c9ec88557910