Address @hides in ImageDecoder
Bug: 152322291
Test: make

Once ImageDecoder is in a mainline module, the system can no longer
access @hidden methods (unless they're @SystemApi), so this
removes/documents some. The ones in Source are not used anywhere outside
ImageDecoder AFAICT, so they are unneeded. Document setOutPaddingRect,
which needs to be public/@hide so it can be accessed by the drawable
package, which will also be in the module.

Change-Id: I5cd9c0553e8d27d6679cbc431fcadc96d818687d

Former-commit-id: 35f16e81dd0b2b4b2f91e79d1102d56063cd3743