Fix layout of DoubleLineTileLayout
With this CL, the layout calculates how many columns it can safely show
given the available space and distributes the tiles accordingly. This
prevents the tiles from going out of the bounds of the container
(happening even in Default Display Size).

With the change in the minimum horizontal margin, the three lowest
Display Size will show three columns and the others will show 2. Showing
two columns has not great expansion animations.

Opening a bug to explore UI.

Fixes: 151810573
Bug: 152429614
Test: manual

Change-Id: I8e123c4497f51f0d5c83b8eb14319f488a96bc9a

Former-commit-id: 6b3f02e2fe8ff88d60d28f454e50a1afc7ca7637