Add atom of TvSettings and its enum
The enums correspond to pages and entries of TvSettings. Some enums that
are related to unreleased features are not included in this CL but will
be added when this CL is cherrypicked into upstream non-public
TV-specific branches.

The atom definition is reviewed and approved by Metrics council.

Metrics council review bug: 151768952
Bug: 150979527
Test: the Android tree still builds normally

Change-Id: I4258312c19aa4f2f0cd178bc5b5dc53e52c6c737
Merged-In: I4258312c19aa4f2f0cd178bc5b5dc53e52c6c737

Former-commit-id: 0bec8d43a11fefe71d1c1659532c0bac5881dfbd