Restore password metrics after a credential-less unlock
Password metrics for the curent lockscreen credential needs to be
available after unlock but it's not possible to recalculate them
if unlock is performed with a token. To fix this we store an
encrypted metrics alongside the current password-based SP and
decrypt it during token-based unlock.

Bug: 145601001
Test: atest com.android.server.locksettings
Test: atest
com.android.cts.devicepolicy.MixedManagedProfileOwnerTest#testResetPasswordWithToken
Change-Id: I01bf310b2e265986be92b09a753f0ecfa69d83f2

Change-Id: Icfc7bd3a3db710e9606cb2df305683b421c785a7

Former-commit-id: 3589e891ba992e2476389f4026e227ddf6989d70