Add ParcelableHolder type
To give a way to extend a stable parcelable object, a wraper for a
parcelable is necessary.
The reason why it needs to be wrapped is allowing to trasfer an extended
parcelable even though some of modules using the parcelable don't
know the type of an extension.

Whenever setParcelable() is called, it checks stability level

Bug: 146611855
Test: atest aidl_integration_test aidl_unittests

Change-Id: I6895976658fd7450b2ee47076a03079995777143

Former-commit-id: a98415b043e505d631126cbb8f163d21eea4461d