Remove window animations for shutdown UI on TV
If reboot happens too quickly, the video output freezes during the
fade in window enter animation. This leads to the text on the shutdown
UI being hard to read, as the window is still transparent and text from
underneath might shine through.
To resolve this, this change removes the window animations from the
shutdown UI for TVs, so that the UI appears instantly without a fade.

Bug: 152509487
Test: flash build and reboot / factory reset
Change-Id: I28153a56f3b5cfc4f89b06fcc8fe6fba935eb30b

Former-commit-id: 2f58649bff1df9662217210ef90ce2d0d036bf1c