Create a shim apk which is malformed
This shim-apk will be embedded in an apex to test out scenarios where
apk-in-apex fails to install.

Since this shim-apk is malformed, PackageManager will fail to parse it.
Hence it should not be possible to install this app at all. Hence I did
not add the hash of this shim-apk inside the manifest of shim-apk-v1
that comes pre-installed with device.

Bug: 152021141
Test: builds succesfully
Test: adb install CtsShimTargetPSdk.apk (fails to parse)
Change-Id: I3509009de62d38c5028185fe3588b674aa006e56

Former-commit-id: 57c68c7482b66e8a47964395866e370dce894b27