Allow PackageDexUsage to record code paths not used by others
The restriction was just an optimization to make the processing simpler
and clearer. All packages are expected to use their code paths and
recording this info was redundant and added noise. However, in
prepartion to record system server code paths we have to record
everything that belongs to a package, even if it's not used by others.

This CL makes PackageDexUsage agnostic of the isUsedByOther flags: it
will record everything that is given to it.

Note that this applies only to PackageDexUsage. DexManager may still
decide to skip the recording (which it does for non system server apps)

Test: atest PackageDexUsageTests DexManagerTests
Bug: 148774920
Change-Id: Icdb9ff45330a8fb15d80cb07495723b473228dda

Former-commit-id: 3f2b728bba995f99b3518713eb4ba3c53644648b