Fix more windows at framework on system gesture mode
When the mode of navigation bar is system gesture mode, there is
a navigation bar inset existing, but the application can not touch
this region due to all touch events from this region received by
launcher. We need to account this region with unaccount space to
avoid the windows which could not be touched passing to A11y
framework.

Bug: 142022890
Test: a11y CTS & unit tests
Change-Id: If87e1f5b840daa93f78c4b4955ed98633b654928

Former-commit-id: fd3b4a8c6940ad48c0137e76d4aa39adacd5720e