Do not use BIND_RESTRICT_ASSOCIATIONS when binding TextClassifierService
The intention was to restrict AiAi, rather than every
TextClassifierService.

Note that AiAi is still restricted because of the allow-association
configs in config.xml.

Test: Patch ag/10873984. Select seme text and then run
"dumpsys activity allowed-associations"
Before the change, extservices is in the output.
After the change, extservices is no longer in the output.

Test: Run a sample app that sends a broadcast to
com.google.android.as/com.google.android.apps.miphone.aiai.reflection.broadcastreceiver.AppInstallBroadcastReceiver
Make sure the association restriction is still in place.

BUG: 152840827

Change-Id: I491be5971e58f967fe3bd6ffdf34126d0644e4fa

Former-commit-id: 2f76d97976ca3f52cd7ffcf3d672370b75f924d8