Compute configuration overrides for tasks created by organizer
The config overrides were only computed for leaf tasks. So, the
config of the organized tasks were inherited from the display,
which was still fullscreen size. This was not right especially
for split-screen organizer tasks because the configuration could
be used by the organizers.

Also remove requesting override windowing mode as fullscreen for
home tasks, the config was updated to fullscreen or
split-screen-secondary while resolving override config instead.

Bug: 152408408
Test: Enter split-screen and show IME
Change-Id: Ie6737cb5e9655c57724fc140c6b6fde3a96be6ab

Former-commit-id: dff902bc7cd073f5db1b3aeb2186528ea1b15021