DeviceConfig: Unban all banned namespaces if any banned one unbanned
There are chances that innocent namespaces got banned with actually
harmful namespace, and since innocent namespaces' flags may not be
changed for a long time, they are getting banned on the devices for a
long time even though they are not problematic.
Thus we want to do: When a namespace's configuration got successfully updated, unban ALL the
banned namespaces(by clearing the hashcode map) if the updated namespace was banned before updated.

Bug: 153010388
Test: atest DeviceConfigTest & local device with Mob Dog
Change-Id: I61b92a09d77ef1e1b7bc9c86f6655ba62a8a3d89

Former-commit-id: f305f8a9f6381bad717ad97abeefc9fe1e8f2e31