Show selected bubble first, after stack expands
When the stack collapses, BubbleData moves the selected bubble
to first in stack, but marks accessed time instead of updated time.

Sorting uses updated time (in repackAll) and moves the selected bubble
back to where it was, when the stack expands later.

Fixed by marking updated time in addition to accessed time.

Fixes: 152883409
Test: manual: select bubble in middle of stack, collapse stack, expand
        stack -> see that bubble first in stack
Test: manual: select bubble in middle of stack, collapse stack, add new
        bubble, expand stack -> see that new bubble is first in stack
        and previously selected bubble is repacked

Change-Id: I8042549708ba3ace13d94e9cd3397af229af2dc9

Former-commit-id: b9be52dc69fff3dd819f3aa94fec69af5657b039