Allow bar to be transparent if letterbox contains bar
When top window request fullscreen and device contain a real cutout
and the layoutInDisplayCutout is default or never, a letterbox will
be created for this activity, which cause
BarController#isTransparentAllowed return false so the status bar
cannot be transparent, the bar will shows a black background
during hidding animation.
For a better animating effect, allow the bar to be transparent if
the letterbox is fully covered the bar.
Also modify the layout area of each letterbox surface, so it can be
more convenient to do geometric check when multiple cutout existing.

Bug: 152273579
Test: manualy verify no black background shows on status bar when
opening fullscreen app.
Test: atest DisplayPolicyTests LetterboxTest

Change-Id: I1f458394a8c6a2ac53a461b6819a026090d62898

Former-commit-id: 39f818c76cf5a2aadee26f3c0905919656a0352b