Add carrier config to indicate if cep is supported on peer.
Adding new carrier config option to indicate if conference event package
data is supported on the conference peer.  We already have an option for
the conference host.  This new option is used to indicate when the peer
can or cannot support the CEP data.  When not supported, even if the
network provides CEP data, it is ignored by the device.

Test: Manual testing using test intents.
Test: Added unit tests for these cases.
Fixes: 150860282
Change-Id: Ib8d8f171699f21f45b0ce2eec2dbe75d7466b76a

Former-commit-id: 60e5f7ca9c30de8bb8ad8b4d91bc478624fa4902