Update bouncer fingerprint lockout logic
Do not request fingerprint auth if the user is locked out and
they have attempted credential on bouncer. This fixes an issue
where the incorrect PIN/Pattern/Password message would be
overwritten by the fingerprint lockout message.

KeyguardUpdateMonitor controlls biometric requests for
lockscreen/bouncer, so poke in credential attempts so it can
be used for determining if biometric should be requested.

Fixes: 141960543

Test: 1) set up PIN/Pattern/Password and fingerprint
      2) reject fingerprint 5 times, then swipe up to bouncer
      3) note that "too many attempts" message is shown
      4) enter wrong PIN/Pattern/Password
      5) note that from now on, this current bouncer session will
         only show credential-related errors

Change-Id: Ifb837b29f93caa5b8762bd5609ca09601b6e4784

Former-commit-id: 8ae64b5c2950eda2c48844e8a798f7b274d8c41a