Check if QQS shows player when updating height
The media session attached to the player may have changed since the last
time the QQS layout was switched. So, check if the QuickQSPanel is showing
a player when deciding how to update the height. Don't check the state
of the media session.

Fixes: 149503712
Test: play podcast, pause and dismiss app. Go to lock screen and swipe
down to see QuickQS. Verify that the panel is appropriately sized.

Change-Id: I1d6252a1ea09d6938fe1b5da6a4bee05b2ebc30c

Former-commit-id: 0dadde727f2aa9bd7e45770624e356a757662a94