Close LS media player when session is destroyed
When updating media player, check the playing state. It is null after
the media session has been destroyed. This is necessary since the
NotificationMediaManager is used as the source of information about the
"current" session. It still sends updates after a media session has been
destroyed.

Fixes: 152821122
Test: play podcast, pause and dismiss app. Go to lock screen. Verify
that LS media player has closed.

Change-Id: I52445d18eb945127f6c9af0908e154ac983e88cf

Former-commit-id: 13bc4269d6c73e2dc51d8d873009ae31049ef674