Move Some Touch Handling out of NSSL
Move some touch handling out of NotificationStackScrollLayout and into
NotificationStackScrollLayoutController.

This corresponds with removing NotificationGutsManager from the NSSL.

The code looks a little funny right now, with the Controller querying
the NSSL for _a lot_ of state that should probably live in the
Controller, but I wanted to keep the size of this change reasonable.

Bug: 147245740
Test: atest SystemUITests && manual
Change-Id: I8ba68ea72898369202b052bee0c46d5c7bd8f032

Former-commit-id: 2cee9ecc9f8697829b1f3c5201fa999d73c02281