Get media devices for each media application
Each player, with a separate pacakge name, needs to register a listener
with its own LocalMediaManager object constructed with that package
name.

An issue with this change is that it assumes that the app package for
the player never changes, which is a bad assumption for the mini player.
Luckily, the mini player doesn't show an output switcher, currently.

Fixes: 152469562
Test: manual -- play demo app and switch output to cast device. Verify
that media player shows name of cast device.
Test: manual -- while casting, play music from another app. Verify that
the media player shows phone speaker while the cast app shows cast
device.

Change-Id: Ica508d1b2f1dc51a28027604418c2a4f8d64aa13

Former-commit-id: 9a7409b68a7c79f437d5045abba57616a8ab7e74