Update the surface size for tasks when config changes
The surface size was only set for the root task and the
surface size was not updated after entering split-screen
mode since those tasks were reparented to split-screen
organized tasks and were no longer be the root tasks.

Applying crop to root tasks, unless the task is created by
organizer. And also remove crop of the descendant tasks.

Bug: 153605363
Test: enter split-screen and rotate device
Test: atest TaskStackTests
Change-Id: I7b21c20dc676aa0be06c3c6e1b6b11c07a3d048d

Former-commit-id: 2570e334250bd8e38250ea2ea9b104dbc67b9c91