Only depend on the latest stable netd AIDL.
Currently, various components in the system server depend
directly on numbered versions of the netd and dnsresolver AIDL
interfaces.

Introduce a new build target that only includes the latest stable
version, and depend on that instead. This ensures that all
components in the system server use the same version of the
generated classes. Depending on multiple versions of the
generated classes in the same (classloader) has in the past
resulted in code compiled using a newer version using a generated
class from an older version, and crashing with NoSuchMethodError
when it calls a method that didn't exist.

Also remove the dependency from wifi code to DNS resolver AIDL
interfaces. Wifi does not (and should not) call the DNS resolver
directly.

Test: m
Bug: 143732914
Bug: 153596226
Change-Id: I2c45853171df8d2c9726f4f12204a48862b30bb6

Former-commit-id: a1c5d9312357b48fed4526b36649084cbc9371b8