Fix notification snooze a11y issues
- Set minimum height and width on the snooze "undo" button
- Update logic in performAccessibilityActionInternal in
ExpandableNotificationRow so that snoozing a notification from
Talkback/switch access/voice access actually performs the snooze action
- After a user selects a new snooze time option for a notification, the
snooze options close. Focus should remain on the snooze time, since that
element was just changed.
- Once a user snoozes a notification, it should focus on the child
SnoozeView instead of the greater NotificationSnooze view because only
the SnoozeView is actionable. So if focus is requested on the
NotificationSnooze view and the options aren't expanded, then focus just
on the SnoozeView.

Test: manual, use a11y scanner
Fixes: 152730944
Fixes: 152820361
Fixes: 152720315
Fixes: 152718852
Change-Id: I7e43bddfdaa3b5a1741645f7bb54af42bc953fc3

Change-Id: I4672654d4ca2eb2cf68539efb7cfb0c8f5dfb00b

Former-commit-id: 90e20d3c6dec163199cd6428894eb3c63ab01d75