docs: Note behavior of PREFER_OFFLINE
Resets developer expectations on the way to deprecation of a parameter
in the Android Speech APIs.

This change was made by hand in the refdoc HTML (cl/303351535);
making the corresponding change to the source code, so the doc
fix won't be reverted with the next DP update.

Bug: 151475334
Test: make ds-docs
Exempt-from-owner-approval: Docs-only change
Change-Id: I053a8e9d9623c394954a118b5e068fc4015b2ba7

Former-commit-id: d6975db81e3599ced4460712774b08bfb42a5cd2