Expand minimized split stack on relaunch-attempt
If a relaunch-attempt is made on the app in a minimized
primary-split, just expand it. This way its clear to the user
what happened and is much closer to their intent. Previously,
this was just opening recents; however, that doesn't make sense
for pixel launcher (which remains in minimized-dock even with
recents active) because the actual launched-app is still
not-focusable and partially obscured.

Also, modified the restart-attempt logic to directly check if
divider is minimized rather than homeTaskVisible since that is
semantically what matters.

Bug: 152772854
Test: Open an app into split-screen primary with launcher in
      secondary. Then launch the same app from launcher.
Change-Id: Ib04d9a949aff5a5202e1f83c820cb3e9589598fb

Former-commit-id: a56c8a2ce2810229eb526c10f9e7af1a537eaddf