Invalidate the cache for APK-in-APEX
PackageCacher uses file name and modified time to determine if the
parsed cache is still valid. However, all APK-in-APEX would have
the same name and modified time. This results in the out-of-date
cache would be treated as valid.

We need to invlidate the cache for the APK-in-APEX when that APEX
is going to be installed.

Bug: 152352677
Test: atest -p frameworks/base/core/java/android/content/pm
Test: atest -p frameworks/base/services/core/java/android/content/pm
Test: Update/rollback an APK-in-APEX, and check its version code

Change-Id: I2a722036d8b1e9c5121d385f8d0667fb908cc7a1

Former-commit-id: 63aa98db6de16abc6a8ef827161ad84e8045f39e