Avoid reporting unhandled config changes to activity
Currently the global configuration is also used as the override
configuration of default display, it is possible that the changed
configuration from relayoutWindow is sent from view root to activity
before handling the relaunch transaction item. So the activity may
receive an unexpected onConfigurationChanged before relaunching.

By enabling the check of activity-handle-able configuration for the
path from view root, the unexpected onConfigurationChanged can be
ignored and proceed to relaunch.

Bug: 145018628
Test: atest ConfigChangeTests#testChangeFontScaleRelaunch
Change-Id: If4c9590583642219ba7b5cd99bbfe84711524cd5

Former-commit-id: e553a89c75bbc5215f085efe7b47bcaf4b0f4c03