Allow app to specify the rollback data policy in the manifest (1/n)
See b/148098864#comment1.

It makes more sense for an app to declare the data policy in their
manifest, rather than leaving this decision to the installer.

The attribute will be read by RollbackManager during install to decide
which data policy to adopt when committing a rollback.

Also migrate to ApkLiteParseUtils per
https://googleplex-android-review.git.corp.google.com/c/platform/frameworks/base/+/11069722/1/core/java/android/content/pm/PackageParser.java#1660.

Bug: 148098864
Test: m
Change-Id: Ie6635249ea0dfbd21b0e56080fb492b9949841ec

Former-commit-id: 68d6ab7815fff610201e3d89d5dc3126241aa1ba