Add min extension versions to AndroidPackage.
The <extension-sdk> manifest tag was added in ag/1004614 to
allow apps to specify the minimum versions they need of
extension SDKs.

This adds the min extension versions to ParsingPackage.

This will be needed to allow us to perform checks that rollbacks do
not violate the min extension versions of installed apps (see
go/sdk-extensions-and-rollback).

Bug: 152737927
Test: atest PackageParserLegacyCoreTest#testUsesSdk
Change-Id: I2e080e46aaea6de489766dac0d323d7d2c7302ca

Former-commit-id: 72ae4476d4a256a15c4b4b4ce9e8eb11b52cd1ab