NotificationMediaManager is bad signal for resumption
Entering resumption on STATE_NONE from NotificationMediaManager
causes flicker because it sends STATE_NONE to often. Instead, we can
listen for PlaybackState changes from the MediaController directly.

For cast, we need to enter the resumption state on STATE_CONNECTING,
because that is the resulting state after ending a cast session.

Fixes: 153937833
Test: manual - play music and toggle play / pause. Look for flicker
caused by action buttons rapidly switching in and out of the
resumption state.
Test: manual - cast from YT to TV. End cast session. Verify that player
is in the resumption state.

Change-Id: Ie7f577d1b06719a1ce2ae5927b2ee20745076303

Former-commit-id: 445d441b441f8d0960d8d8c9aa56a7fa6c400ff6