Allow WindowOrganizerController to handle DisplayAreas
The current code for WindowOrganizerController only allows transactions
to be submitted for Tasks. This change makes the request handing more
generic so it should allow for transactions to be submitted for all
WindwoContainers. However, we're limiting the requests to Tasks and
DisplayAreas to avoid any larger issues.

There are some transactions that will only work for Tasks so there
is a separate code path to handle Task specific transactions.

Test: WindowOrganizerTests
Bug: 152116619
Bug: 152114574
Change-Id: I916a8f0c92be490318ee45e09f6a5e8a4ee6ec62

Former-commit-id: 6630d85d7cfd32b76389d546438a8342501ab99a