Ensure ADB is stopped when enabling Test Harness Mode
Because adbd may be enabled before TestHarnessModeService
activates Test Harness Mode, we need to stop ADB by turning the
setting off. For user builds, this is a no-op. For userdebug and
eng builds, this stops adbd and restarts it immediately afterwards.
This allows the newly-created keys to be loaded when adbd starts
up.

Change-Id: Ia673d1c18aba2e33d4df901ffe982122da4cb295
Bug: 150721793

Former-commit-id: e2f1fe31e7f94c26e3c9208d061b11cac1a63655