Record supported extension versions at rollback creation.
This adds a property to the Rollback class to record which extension
versions were supported at the time of creation.

This will be used to block commit of a rollback if it may result in
an installed app's minExtensionVersion being violated. See
go/sdk-extensions-and-rollback for details.

Bug: 152737927
Test: atest RollbackStoreTest
Test: atest RollbackTest
Merged-In: If770a615a1b23505477d786d6ab204e32d9af4c5
Change-Id: If770a615a1b23505477d786d6ab204e32d9af4c5

Former-commit-id: 2986572ae8b84a4607af2cc123c163cef173b38f