A11y services can perform click on keyguard notifs
When notifications are showing on the keyguard, notifications must be
double-tapped to open. These touches are intercepted in the
ActivatableNotificationViewController's touch handler, so we do not need
to double-check for mNeedsDimming on performClick in the view. The
removal of this check allows a11y services to perform clicks on the
view.

Test: manual
Fixes: 130588833
Change-Id: Ie4b9a928599a056dba982ef12f974b8a2bf14f59

Former-commit-id: d075b63c5287281796b778f8239c41d03f14691a