Copy leashes about to return to prevent others release them
When a thread is excuting addWindow or relayoutWindow, there might be
another thread attempting to call updateInputMethodTargetWindow but
waiting for the lock. When the lock is released by addwindow or
relayoutWindow, the logic inside updateInputMethodTargetWindow might
release the leash that addWindow or relayoutWindow is about to return.

This CL return the copied leashes, so that other thread won't release
the leash which is about to return.

Fix: 154175855
Test: atest CtsWidgetTestCases
Change-Id: I48a7dbf1576e07302c66955e2d3c86037e987617

Former-commit-id: 9629242dd4b19c974e80cb772eaf7af6d79fae38