InputMonitor: Ensure we assign input to correct Surface
We want to assign input to the BLAST surface instead of the
WSA container layer if BLAST is enabled. Previously this didn't
matter but following upcoming occlusion changes the BLAST surface
would then occlude the WSA surface and the main Input Channel would
fail to receive input. We fix this by having InputMonitor call
getClientViewRootSurface()

Bug: 152064592
Test: Flip BLAST. Input still works!
Change-Id: I3e4e136fa11d15f8f576b97dbdc34c9a585a0f32

Former-commit-id: e57724d7fcd58229a8a61ec796eb73a413bdc56c