Update ProcessRecord ApplicationInfo for RROs
When overlays update, the changes in ApplicationInfo are propagated
to affected processes. The ApplicationInfo in the ProcessRecords of the
running processes are not updated in ActivityManagerService. Persistent
processes reuse the existing ProcessRecord when they are restarted, so
they are restarted with outdated overlay paths.

This change updates the ApplicationInfo of ProcessRecords when overlay
paths change.

Bug: 154242963
Test: open settings > system > gestures
      switch to another navigation mode
      kill sysui (adb shell kill `pid com.android.systemui`)
      Confirm that navigation mode has not been reset
Change-Id: I85f89df538079e1fb1d695a9b13b388dc338b47b

Former-commit-id: 03bf28bf5c62deb2e1d420a0a9028a09412af972