Extend isModule check for apk-in-apex
If an apk inside an apex is the failing package in a rollback,
we do not retrieve the logging parent of that apk since it is not
considered a module by PackageManager. Instead, extend the logic
to retrieve the logging parent of a package that is an apk-in-apex.
This is done by querying the existing ApexManager mapping.

Test: Manual. Test logging flow on
com.google.android.providers.media.module before and after patch
Bug: 154360162

Change-Id: I264bcd1beff89c966b5733f6be713166d066fc46

Former-commit-id: 430d5fb8172fd9ad0fc5794ee1b29036967c06d3