Prevent KeyguardContainer size from shrinking when switching from Password Security Challenge.
Before this change, KeyguardPasswordView used maximum height and width.
First, since the Bouncer is the only component in Car Keyguard, no such
configuration was needed.
Second, once the Keyguard Container was shrunk in size due to the max
height and width config of KeyguardPasswordView, it did not increase
again to host the view of other Security challenges such as PIN and Pattern.

As a result, the PIN and Pattern security challenges were cropped out
due to insufficient space.

Therefore, the max height and width attributes of Keyguard Password View
has been removed.

Test: Manual
Bug: 154555720
Change-Id: I56b38d5401c7c88fdfedf33f11f713217f6dbc4d

Former-commit-id: 6042626a97cfd7bfcbb40eab5b49ed407dfc6424