Restore share sheet height when toggling dark mode
This CL does the following changes:
- Save and restore the resolver drawer layout height.
This is necessary because when we toggle darkmode,
we don't go through handleLayoutChange, thus we don't
recalculate the height the standard way.
- Disables saving state for the viewpager.
This is necessary, because the content is
incorrectly restored afterwards. It's better to
reload it instead.
- Saves and restores the last selected tab. This is
necessary because with the viewpager state saving
disabled, it defaults to the first tab.

I considered updating the configChanges attribute in
the manifest to include dark mode toggling, but it
gave unsatisfactory result - the share sheet only
partially became dark.

Test: manual
Test: atest ChooserActivityTest
Test: atest ResolverActivityTest
Fixes: 154526324
Change-Id: I99e159ea7ca413f8acbc0429a43ce5aba32a1b28

Former-commit-id: 9d8941457dbdaf1ac874307aa137b539fddc5cb0