Use WC bounds for insets calculation
When computing insets based on navigation bar and status bar, use the
window's bounds instead of display bounds to ensure it compenstates for
the display area. The system bars should use their bounds to ensure
they aren't creating insets outside the display area. The window bounds
are also used when calculating the size and placement of the system
bars.

Bug: 156676090
Test: DisplayPolicyLayoutTests
Test: DisplayPolicyInsetsTests
Test: InsetsPolicyTest
Test: InsetsSourceProviderTest
Test: InsetsStateControllerTest
Change-Id: I9a3481fcd0146463c10077967b49bf46190f7946

Former-commit-id: c9889ea63c37dc107d87977c03faba91e5dfa8ed