Ignore task-info changed callbacks in uninitialized Divider
Apparently there is a chance that SysUI crashes without
system server receiving binder-death (or maybe it receives
binder-death too late?). In this situation, when sys-ui
restarts, the split root-tasks will still exist so there is
a small window of time during init() where WM might send
messages to the new organizer before init() finishes
running.

This avoids a cycle of crashes by returning early, but if
this is hit, the system is likely in a weird state and
Divider probably won't work.

Bug: 153811023
Test: Will need to wait and see. Too low repro rate.
Change-Id: Id42569b4699699dbc7eaa63eaff801c7c52314c7

Former-commit-id: a5858ed4e44fbc1b6c291d2e7506952f95633b06