Add multi-user support for AutoTileManager
Before this CL, AutoTileManager (and AutoAddTracker) would only work
"corectly" for user 0, as all the settings (read and write) were done in
user 0. This meant that tiles that got auto added as user 0 would not be
auto-added for other users. Similarly, tiles that got auto added as a
secondary user would not be auto added as user 0.

This CL makes sure that everything is done with respect to the current
user (as reported by QSTileHost).

Test: manual
Test: atest AutoTileManagerTest AutoAddTrackerTest
Fixes: 154607890

Change-Id: I90717dedc8935dd2597e504acd57f5b44236ef3e
Merged-In: I518334cca0a0960e0b3b74b43ac9fce4610c7bd0

Former-commit-id: 29d65f5e3d5f170d350f7c59c821994b96b0e6d6