Add multi-user support for AutoTileManager
Before this CL, AutoTileManager (and AutoAddTracker) would only work
"corectly" for user 0, as all the settings (read and write) were done in
user 0. This meant that tiles that got auto added as user 0 would not be
auto-added for other users. Similarly, tiles that got auto added as a
secondary user would not be auto added as user 0.

This CL makes sure that everything is done with respect to the current
user (as reported by QSTileHost).

Checks if the color display supports inversion in tests.

Test: manual
Test: atest AutoTileManagerTest AutoAddTrackerTest
Fixes: 154607890

Merged-In: I518334cca0a0960e0b3b74b43ac9fce4610c7bd0
Change-Id: Icc1406ebd2bf94da3a427f3b13f11b325e25b491

Former-commit-id: 61ebaf5c46220852b6bae9b19db75cb42a21ed6d