Do not schedule transition in onTaskVanished
With the latest change, dismiss a PiP should now be originated from
SysUI itself. When the onTaskVanished callback happens as a result of
removeStacksInWindowingMode, there's no necessity for SysUI to start
transition again.

Plus, the SurfaceControl would be revoked when onTaskVanished happens
and it would be a fatal error if SysUI does so.

Bug: 153280613
Bug: 152214447
Test: manually enter/exit/close PiP
Change-Id: I785eaac52fa1410ef2a22c41035cf952410a7335

Former-commit-id: 180162e1e12a31abf6c212a88d04585c8533572c