Add logging to Dream stop.
Unexpected stop in dreams get logged very generically and are difficult
to debug. Additionally, in many cases of stopping, DreamController
requests the dream to stop wherein the dream finishes itself.
This ultimately gets logged as a normal finish. Added code to save
the context of the original request to stop so that it gets logged
appropriately.

Bug: 150436621
Test: Manually tested the new logged reasons.
Change-Id: I25e348026b514dffe540cf8c4cf3a9b87d385fde

Former-commit-id: 9c988b2980a3f8c6dcb83e6ca7812fdf7d823f13