Handle variable class loader context for system server
Do not dexopt system server dex files which have a
variable class-loader context. We could extract them
but that will interfer with ZygoteInit code which will
try to verify them again.

As part of this change, also improve PackageDexOptimizer
error handling, logging wtf message when we cannot retrieve
the dexoptStatus.

Test: manual
Bug: 154789494
Change-Id: I8a8181ce637b13e5cab10267c50c0dccb187b15b

Former-commit-id: 1ef2b2b6eba29af497cc0d7b03d192cd8a2c1776