Fix test failure in testProcessStateByPulling
Added a new flag "--statsd" to dumpsys procstats, it'll dump
the identical protobuf data as the one being sent to statsd;
these data is aggregated/reduced.  The tradtional "--proto"
is still supported in case the full data is needed.

Align the ProcStats's proto message definition with the statsd.

Fixed various other issues with ProcStats's dumping.

Bug: 148542701
Test: atest ProcStatsValidationTests
Change-Id: I5a22603bfbc97bfac93179289df839710364677d

Former-commit-id: 468e6c390958e524bac54be367611cc19569a0b9