Fixes NON_BROWSER flag behavior
This change modifies the time we do filtering / blocking based on the
non browser flag. Before this we would filter out browsers from results
in the component resolver, leading to the potential to ignore user
preferences. With this change, we apply user preferences and only fail
the start if the single resolve info or all potential targets sent to
the resolver activity are browsers.

Test: atest MatchFlagTests
Fixes: 154766920
Change-Id: I0451eb2e199342d369851294ebaf4affb5fe86c6

Former-commit-id: b3ab3b1c05925a7a27fc2ea41227e1c398bcbb38