Keep screenshot process bound
Currently, the screenshot process only remains bound to SysUI
(by ScreenshotHelper) until the screenshot finishes saving. This
makes it vulnerable to getting cached and frozen. Since the process
now includes UI with touch focus for longer than that, this can
cause ANRs in the screenshot process, if it is frozen while it's
supposed to be accepting touch input.

Retaining the binding for the entire lifecycle of the screenshot
UI fixes this problem.

Test: manual -- used the steps in b/153577093 to verify the
freezing behavior and verified that it no longer occurred after
the change
Bug: 153418099
Fix: 153418099

Change-Id: Idca24a69274af3d633e1243b46b602613cb04d50

Former-commit-id: a4c79758fb53a6d25368c9fe706c666a7d6f5db3