Add back ActivityManagerService.onUserRemoved
ActivityManagerService.onUserRemoved was accidentally removed in
ag/4246926. This cl adds it back.

The part of onUserRemoved that told ActivityTaskManagerService that
the user had stopped (and thereby cleaned up ATMS) was retained, but
the additional actions of AMS.onUserRemoved - namely, telling
BatteryStatsService and UserController - were accidentally removed.
Consequently, those services were never told when a user was removed.

Bug: 140645064
Test: manual verification that profile is removed in  adb shell dumpsys activity users
Change-Id: I903b06ae68d583bdb326ad3b764a63d18b5555d8

Former-commit-id: aa59419bbdacebd6e13159636391ad541a09df61