Fix lock icon animation for bio-auth.
1) Ensure lock icon animates when using biometric authentication to
unlock a device.

2) Don't animate the lock twice. Accomplished by waiting to clear
the biometric authentication credentials until after we've started
the keyguard dismiss animation. When we cleared them before the
animation, the keyguard would be left in an inconsistent state where
it would be non-dismissable but also not-yet animating away, meaning
the keyguard appeared locked to the rest of the codebase.

Fixes: 153117689
Test: atest SystemUITests && manual
Change-Id: I8c2390e5f3ca2d366d377f6da5822f24ca311c2f

Former-commit-id: 27e34b2a0331e891d998730db4c2405cdde7ea1b