Fixes StaticSharedLib state issues
This change fixes a few issues with the way static libs were marked as
installed. Primarily, it marks any static shared lib as installed for
any user that has a dependent package of it also installed. This ensures
we never end up in a situation where an app is installed but its static
shared lib is not.

Test: atest StaticSharedLibsHostTests
Fixes: 143701280
Change-Id: Ia66da8e54162f53a236d300674b3c634cf008d05

Former-commit-id: c2e24b1c9a170291c93c2fa5dd9bda77b89951df