Shift Primary stack as well as secondary during ime adjustment
Previously, was only 'adjusting' the secondary split for
ime. So the secondary had functional insets and drag/drop.
However, the primary wasn't being adjusted (accorcding to
WM). This caused its insets to be calculated against its
originla position (which allowed ime insets to potentially
overlap it) and it meant that WM didn't use its visual
position when calculating drag/drop coordinates.

So, this does the same with the primary stack as what was
done with secondary.

However, doing this required some other fixes in WM. Basically,
if appbounds is explicitly overridden, just use them instead of
intersecting with parent. Also, WSA doesn't apply crop for
splits (maybe it shouldn't apply crop to anything anymore).

Additionally, insets calculation was applying top insets to the
bottom if the inset frame's top wasn't exactly equal to the
window's top -- because of a catch-all else condition. So this
adds checks for matching the bottoms as well.

Bug: 154056542
Bug: 151862790
Test: Open two apps in split screen and open IME in secondary. Drag
      from secondary to primary.
Change-Id: I2391783e726e4a1c8ed3150628af2f398218fe90

Former-commit-id: e067ddfc05c495eebf7e07664774a0776ec7813d