Only perform autolaunch if list has completely rebuilt
Previously we would check for autolaunch the first time
onPostListReady is called, but that is too early in the case
when we start the async task for sorting the targets. In
that case onPostListReady is called twice - once before
task is launched, and once after task is finished.

We fix this by passing the rebuildListCompleted parameter to
the onPostListReady callback and use it to check for autolaunch.

Test: atest ChooserActivity
Test: atest ResolverActivity
Test: manual
Fixes: 154728817
Change-Id: Ib1b049af00d791898b908bdb65971c07c558f3ac

Former-commit-id: 1b11cc1a17de59f284b1b3dc28375ac3bc63c767