Give Divider back its touch-region
Divider used to have a touch-region that allowed apps
to receive touch events that would normally overlap the
divider window as long as they weren't in the center.

Re-introduce this by allowing windowless windows to
pass through a touch-region. This region is always
constrained to the window size to prevent security issues.

Bug: 154840035
Test: Enter split and observe touch events away from
      center of divider.
Change-Id: I85310c78beda7bd5574df78e2a83462e965d1154

Former-commit-id: 128372810ddc47891b5cb93f6a61844bb474e425