Use the rollback data policy specified in the manifest (2/n)
Now app can specifies the rollback data policy in the manifest.

Note the policy specified in the manifest will take precedence only when
it is not the default (i.e. RESTORE) in order not to break the existing
code. When Play has done migration and stopped using
SessionParams#setEnableRollback(boolean, int), we will remove related
code and manifest will be the single source of truth to specify the
rollback data policy.

Bug: 148098864
Test: atest RollbackTest StagedRollbackTest
Change-Id: Ia8ac9d710b7c36519672b4a90eccea4f6836b5c5

Former-commit-id: aad90de1e0e9bbecc08610d1075d19f2eac334ed