Update when media controls get cleared
Some apps include an action to dismiss a media notification, so we
should listen for that happening and clear controls in that case.

Also, remove STATE_CONNECTING as a condition to clear controls -
This was originally added in ag/11056932 as a workaround for an issue
with YouTube cast sessions. However this caused issues with other apps
like Spotify which set STATE_CONNECTING while still active. YT was using
that as a workaround for legacy behavior and will update to use
STATE_NONE for R+ builds (b/155213698). In the meantime, listening for
when the notification is removed will also work to clear YT's controls
as expected.

Fixes: 154953276
Fixes: 155029855
Test: manual

Change-Id: Ie9320e1406c1f457a39f67705ec1ffcb3a983488

Former-commit-id: d664de2cd5eba8d95092b4ea68540e7f91c2081f