Highlight provided channels
The ChannelEditorDialogController is handed up to 4 channels from
NotificationInfo when presented, and pads out to 4 in the case when it
gets fewer than 4 (the most common case is that it gets 1).

This change tells the controller to play a highlight animation for the
provided channel(s) so that it's more obvious to users which channel
they are editing the preference of.

Test: atest SystemUITests; manual
Bug: 140956160
Change-Id: I48717c7ca6e54259e54f32a5af6acc4c264cc8eb

Former-commit-id: 0fd065412908ede47512bbc80b3b2945479cada7