Add mBouncer null checks for view-related requests.
CarKeytuardView is inflated when CarKeyguardView is first started, and
mBouncer is assigned a value onFinishInflated.

Since some of the view-related requests that reference mBouncer could be
called before that event, we need to add null checks for them.

Such a check is redundant for methods that check for mShowing, since if
mShowing is true, it is guaranteed that mBouncer is assigned a value.

Bug: 154928953
Test: Manual and Existing Unit Tests
Change-Id: I796e8bc73459df3b2a0374818556262ebc2ea237

Former-commit-id: 16e0cbc048c575ee60f9944127797690b2b2e15e