Use ConstraintLayout for screenshot UI
There are some problems with the screenshot UI layout, especially
in RTL (where the dismiss button disappears). Using constraints
rather than setting the position of the button directly fixes this
problem, and should make updates to the appearance of the UI in
landscape simpler.

Bug: 151814483
Fix: 151814483
Test: manual -- made sure that the animation and final UI state
looked correct in both LTR and RTL layouts

Change-Id: Ia8d21c028279bc6b70a1b338bea979f2ce2808bc

Former-commit-id: 9dd850f43d8f3e07cc7c692351d827ac06e1fbe3