Don't throw exception for inflation after removal
There are some cases where we seem to be inflating after the removal of
a notification, most likely related to a re-entrant call where we
remove a notification as part of a notification update listener.  This
throws an exception since we assume a removed notification doesn't need
to inflate anything.

Until we fix the underlying issue, we loosen the guarantee and remove
the exception and instead just return early.

Bug: 155324756
Test: builds
Change-Id: I702391311e9d06319f4d93dd584ecc2a4dd78a36

Former-commit-id: a61906457c647a49cda34aafebcff331b681f1a5