Prevent double animation of lockicon when waking.
If the phone went from pulsing -> awake, it could cause the lock icon
to briefly disappear then reappear. Specifically, this caused it to
appear to animate twice.

This all resulted from the internals of the Doze system briefly being
told that it was neither dozing nor pusling, meaning that it thought
it should hide the lock icon. The rest of the system then caught up,
and it shows the lock icon again, animating it into place.

Fixes: 155411884
Test: manual
Change-Id: I79e1bbde2c5cb5fe588a29111294fab68c546c7c
Merged-In: I79e1bbde2c5cb5fe588a29111294fab68c546c7c

Former-commit-id: 50883d72aa7b32faf697737e650e2677a35125d6