Call finishDrawing for Windows with visiblity GONE
The viewVisiblity is set to GONE with a client request to relayout. If
this occurs and there's a blast sync transaction waiting, finishDrawing
will never be called since the client will not render when visibility is
GONE. Thereofre, call finishDrawing here to prevent system server from
blocking on a Window that will never draw.

Change-Id: I04f143c962dd94e6ba0e439d421eb3a53a6548cc
Bug: 153579514
Test: applySyncTransaction with hiding window in transaction

Former-commit-id: 0cf6852761990d2ae39f8453865d26048774627e