Add overlayable configurator resources
This change adds string/config_overlayableConfigurator and
array/config_overlayableConfiguratorTargets.

string/config_overlayableConfigurator represent the package name that
is recognized as an actor for the packages listed in
@array/config_overlayableConfiguratorTargets. If an overlay targeting
one of the listed targets is signed with the same signature as the
configurator, the overlay will be granted the "actor" policy.

This change does not affect package visibility nor does it grant the
configurator permission to enable/disable overlays targeting the
listed target packages.

Bug: 154610560
Test: atest OverlayManagerServiceImplTests
Change-Id: Id4a803ec179bfeba1b4c5e14bedd18f4cd5d5b23

Former-commit-id: 639393e1e6bd87b21bab89b30c3da0f1431d80a1