Fix logic inversion bug for "basicGrant".
Basic grants are defined to be one that only involve read/write
permissions, which we attempted to validate by checking for any other
flags, but there are many other (completely valid) flags that are
unrelated to Uri permissions.

Thus we invert the logic to look explicitly for the two specific
modes that cause a grant to be non-basic: persistable and prefix.

Update tests to verify the fix.

Bug: 155717493
Test: atest FrameworksServicesTests:com.android.server.uri
Change-Id: Ic8d7980d99bba59fb389d34f3bd054a9907df55c

Former-commit-id: 51a14c09bcf981dbd1a04ca8eab69b1e8e0799d7