Remove redundant primary-stack movement in start-from-recents
The createStack routines used by startActivity already handle
the case of converting WINDOWING_MODE_PRIMARY into move-to-
primary root-task. Doing that work in startActivityFromRecents
was redundant.

Additionally, because the recents version was executing before
the normal activity-start logic, it was making the new
activity visible before postStartActivityProcessing which is
supposed to run before the activity-start happens. It was
effectively starting the activity twice which is wrong and
also wasteful.

Bug: 154575664
Test: Alternate launch activity from recents into primary and
      expand it to fullscreen repeatedly.
Change-Id: I5c6341f8589e8ac3ddd4afb728efd05bd9445945

Former-commit-id: 89cfd63c1946b7444c0b897a9ff77e992b4db01b