Skip registering callback for appop changes if fuse is enabled.
The work done in StorageManagerService when opChanged() callback
gets triggered is causing delay in sending onLockedBootCompleted()
callback during the user switch. Given that opChanged() computes
the mount mode but does not use it for anything, this change skips
listening to these appop changes to speed up the user switch
process.

Bug: 134624284
Test: atest cts/hostsidetests/appsecurity/src/android/appsecurity/cts/ExternalStorageHostTest.java
      (Same P/F with and without this change)
Change-Id: I4bbdfa7954ab0499e4828f4bff0b2ce80444471b

Former-commit-id: 3567f2be56a1b27548ba9048d3b945e9d03d8323