Remove instances of @hide from RecordingCanvas
Bug: 155905258
Test: make

Remove MAX_BITMAP_SIZE. It was based on a limitation of the old
renderer. Remove override of throwIfCannotDraw, which was only there for
comparing the size of a Bitmap against MAX_BITMAP_SIZE and throwing if
it was too big.

Remove MAX_BITMAP_SIZE from WallpaperManagerService, which was a copy of
the one in RecordingCanvas, and only there due to it.

Mark obtain, recycle, and finishRecording as being used by the package.

Change-Id: I141d6d79520871df1a3c2a9267ec872a8314ffdb

Former-commit-id: 7648ccaa3b2ff2f058bd9909d1ecc1e8049bc5be